Use the local proxy/registry
Part of the assumed infrastructure for the IRIS workshop is a proxy host which
runs a web server, Docker registry mirror, and a caching web proxy.

We use the registry mirror to pull Docker images, the web server to host CentOS
cloud images, and the proxy for RPM packages.

It should be possible to use the proxy for other things, but this requires
support in Kayobe for setting the http[s]?_proxy environment variables
appropriately during image downloads etc. The proxy also cannot cache HTTPS
requests, so we need to use HTTP URLs with appropriate verification of
downloads via checksums etc.